Added export.py naming convention for all equipment and subsequent categories
