+++ wizard-wars/lib/rlengine/equipment/armor/boots/boot.py
@@ -1,0 +1,6 @@
+from ..item import *
+from random import randint
+
+class Boots(Armor):
+    def defaultType(self):
+        return 'boots'
