+++ wizard-wars/lib/rlengine/equipment/weapons/fangdagger.py
@@ -1,0 +1,6 @@
+from ..item import *
+from random import randint
+
+class FangDagger(Weapon):
+    def specialMeleeAttack(self):
+        return ['venom', {'power': 4, 'owner': self.owner}]
