+++ wizard-wars/lib/rlengine/items/items.py
@@ -1,4 +1,4 @@
-from equipment import *
+from equipment.export import *
 from potions.potions import *
 from tools import *
 from wands.wands import *
